+++ api/src/modules/scenarios/dto/create.scenario.dto.ts
@@ -1,4 +1,13 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
+import {
+  IsInt,
+  IsOptional,
+  IsString,
+  IsUUID,
+  Length,
+  Max,
+  Min,
+} from 'class-validator';
 import { Dictionary } from 'lodash';
 import { ScenarioType } from '../scenario.api.entity';
 
@@ -7,7 +16,8 @@ export class CreateScenarioDTO {
   name: string;
 
   @ApiPropertyOptional()
-  description: string;
+  @IsOptional()
+  description?: string;
 
   @ApiProperty()
   type: ScenarioType;
@@ -15,6 +25,29 @@ export class CreateScenarioDTO {
   @ApiProperty()
   projectId: string;
 
+  @ApiProperty()
+  @IsString()
+  @Length(3)
+  country: string;
+
+  @ApiProperty()
+  extent: Record<string, unknown>;
+
   @ApiPropertyOptional()
-  metadata: Dictionary<string>;
+  @IsOptional()
+  wdpaFilter?: Record<string, unknown>;
+
+  @ApiPropertyOptional()
+  @IsInt()
+  @IsOptional()
+  @Min(0)
+  @Max(100)
+  wdpaThreshold?: number;
+
+  @ApiProperty()
+  @IsUUID()
+  adminRegionId: string;
+
+  @ApiPropertyOptional()
+  metadata?: Dictionary<string>;
 }