+++ app/layout/header/__test__.js
@@ -1,5 +1,0 @@
-context('Header', () => {
-  it('should have logo', () => {
-    expect(true).to.be.equal(true);
-  });
-});